Moves signing determinism tests to `test-signing` and adds non-exportable tests
